fix file naming convention
