+++ operating_point.py
